+++ plugins/fr.cea.nabla.ui/examples/NablaExamples/src-gen-cpp/stl-thread/explicitheatequation/ExplicitHeatEquation.h
@@ -41,12 +41,12 @@ public:
 		double Y_LENGTH;
 		double u0;
 		RealArray1D<2> vectOne;
-		size_t X_EDGE_ELEMS;
-		size_t Y_EDGE_ELEMS;
+		int X_EDGE_ELEMS;
+		int Y_EDGE_ELEMS;
 		double X_EDGE_LENGTH;
 		double Y_EDGE_LENGTH;
 		double option_stoptime;
-		size_t option_max_iterations;
+		int option_max_iterations;
 
 		Options(const std::string& fileName);
 	};
